+++ emugui.py
@@ -56,6 +56,7 @@ from dialogExecution.editVMNew import EditVMNewDialog
 from dialogExecution.win81NearEOS import Win812012R2NearEOS
 from dialogExecution.errDialog import ErrDialog
 from dialogExecution.settingsRequireRestart import *
+from dialogExecution.startVMNew import StartVmNewDialog
 
 try:
     import translations.de
@@ -110,6 +111,8 @@ except:
     print("EmuGUI has to warn you.")
     print("Error code: W-12-JBJM9")
     print("If this error occurs multiple times, contact your administrator and/or ask for help on the EmuGUI Discord Server or on its GitHub repository.")
+
+import services.vm_data as vmd
 
 class Window(QMainWindow, Ui_MainWindow):
     def __init__(self, parent=None):
@@ -1690,6 +1693,7 @@ class Window(QMainWindow, Ui_MainWindow):
         cursor = connection.cursor()
         logman = errors.logman.LogMan()
         logman.logFile = logman.setLogFile()
+        
 
         try:
             cursor.execute(debug_db_settings)
@@ -1704,7 +1708,7 @@ class Window(QMainWindow, Ui_MainWindow):
             get_vm_to_start = f"""
             SELECT architecture, machine, cpu, ram, hda, vga, net, usbtablet, win2k, dirbios, additionalargs, sound, linuxkernel,
             linuxinitrid, linuxcmd, mousetype, cores, filebios, keyboardtype, usbsupport, usbcontroller, kbdtype, acceltype,
-            storagecontrollercd1, storagecontrollercd2, hdacontrol, cd1, cd2, floppy
+            storagecontrollercd1, storagecontrollercd2, hdacontrol, cd1, cd2, floppy, timemgr, bootfrom
             FROM virtualmachines WHERE name = '{selectedVM}'
             """
 
@@ -1714,33 +1718,19 @@ class Window(QMainWindow, Ui_MainWindow):
                 result = cursor.fetchall()
 
                 print(result)
+                
+                vmdata = vmd.VirtualMachineData(
+                    selectedVM, result[0][0], result[0][1], result[0][2], result[0][16], result[0][3], result[0][5], result[0][6],
+                    result[0][9], result[0][17], result[0][11], result[0][12], result[0][13], result[0][14], result[0][15], result[0][18],
+                    result[0][21], result[0][19], result[0][20], result[0][22], result[0][23], result[0][24], result[0][25], result[0][26],
+                    result[0][27], result[0][28], result[0][29], result[0][30], result[0][4], result[0][10]
+                    )
 
                 architecture_of_vm = result[0][0]
-                machine_of_vm = result[0][1]
                 cpu_of_vm = result[0][2]
-                ram_of_vm = result[0][3]
-                hda_of_vm = result[0][4]
-                vga_of_vm = result[0][5]
-                net_of_vm = result[0][6]
                 usbtablet_wanted = result[0][7]
                 os_is_win2k = result[0][8]
-                dir_bios = result[0][9]
-                additional_arguments = result[0][10]
-                sound_card = result[0][11]
-                linux_kernel = result[0][12]
-                linux_initrid = result[0][13]
-                linux_cmd = result[0][14]
-                mouse_type = result[0][15]
-                cpu_cores = result[0][16]
-                file_bios = result[0][17]
-                kbd_type = result[0][18]
-                usb_support = result[0][19]
-                usb_controller = result[0][20]
-                kbd_layout = result[0][21]
                 accel_type = result[0][22]
-                cd_control1 = result[0][23]
-                cd_control2 = result[0][24]
-                hda_control = result[0][25]
 
             except sqlite3.Error as e:
                 print(f"The SQLite module encountered an error: {e}.")
@@ -1760,60 +1750,7 @@ class Window(QMainWindow, Ui_MainWindow):
 
                 dialog = ErrDialog(self)
                 dialog.exec()
-
-            if platform.system() == "Windows":
-                tempVmDef = platformSpecific.windowsSpecific.windowsTempVmStarterFile()
-        
-            else:
-                tempVmDef = platformSpecific.unixSpecific.unixTempVmStarterFile()
                 
-            try:
-                with open(tempVmDef, "w+") as tempVmDefFile:
-                    tempVmDefFile.write(selectedVM + "\n")
-                    tempVmDefFile.write(architecture_of_vm + "\n")
-                    tempVmDefFile.write(machine_of_vm + "\n")
-                    tempVmDefFile.write(cpu_of_vm + "\n")
-                    tempVmDefFile.write(str(ram_of_vm) + "\n")
-                    tempVmDefFile.write(hda_of_vm + "\n")
-                    tempVmDefFile.write(vga_of_vm + "\n")
-                    tempVmDefFile.write(net_of_vm + "\n")
-                    tempVmDefFile.write(str(usbtablet_wanted) + "\n")
-                    tempVmDefFile.write(str(os_is_win2k) + "\n")
-                    tempVmDefFile.write(dir_bios + "\n")
-                    tempVmDefFile.write(additional_arguments + "\n")
-                    tempVmDefFile.write(sound_card + "\n")
-                    tempVmDefFile.write(linux_kernel + "\n")
-                    tempVmDefFile.write(linux_initrid + "\n")
-                    tempVmDefFile.write(linux_cmd + "\n")
-                    tempVmDefFile.write(mouse_type + "\n")
-                    tempVmDefFile.write(str(cpu_cores) + "\n")
-                    tempVmDefFile.write(str(file_bios) + "\n")
-                    tempVmDefFile.write(kbd_type + "\n")
-                    tempVmDefFile.write(str(usb_support) + "\n")
-                    tempVmDefFile.write(usb_controller + "\n")
-                    tempVmDefFile.write(kbd_layout + "\n")
-                    tempVmDefFile.write(accel_type + "\n")
-                    tempVmDefFile.write(cd_control1 + "\n")
-                    tempVmDefFile.write(cd_control2 + "\n")
-                    tempVmDefFile.write(hda_control + "\n")
-
-            except:
-                if platform.system() == "Windows":
-                    errorFile = platformSpecific.windowsSpecific.windowsErrorFile()
-        
-                else:
-                    errorFile = platformSpecific.unixSpecific.unixErrorFile()
-
-                with open(errorFile, "w+") as errCodeFile:
-                    errCodeFile.write(errors.errCodes.errCodes[36])
-
-                logman.writeToLogFile(
-                    f"{errors.errCodes.errCodes[36]}: Could not write VM data onto temporary definition file."
-                    )
-
-                dialog = ErrDialog(self)
-                dialog.exec()
-
             if usbtablet_wanted == 1:
                 if platform.system() == "Windows":
                     errorFile = platformSpecific.windowsSpecific.windowsErrorFile()
@@ -1877,7 +1814,7 @@ class Window(QMainWindow, Ui_MainWindow):
                     for res in result_settings:
                         if res[0] == f"qemu-system-{architecture_of_vm}":
                             if res[1] != "":
-                                dialog = StartVirtualMachineDialog(self)
+                                dialog = StartVmNewDialog(vmdata, self)
                                 dialog.exec()
                                 arch_supported = True
                                 break